Made global repository routes network, local, database repositories.
